requirements.txt
# ============================================================================
# Global_Searching - Requirements
# Diffusion Model Inference with Local and Global Search
# ============================================================================

# ============================================================================
# Core Dependencies
# ============================================================================
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0
pillow>=9.5.0
tqdm>=4.65.0

# ============================================================================
# Configuration Management
# ============================================================================
pyyaml>=6.0
omegaconf>=2.3.0

# ============================================================================
# Evaluation Metrics
# ============================================================================
# For FID calculation - choose one:
clean-fid>=0.1.35
# Alternative: pytorch-fid>=0.3.0  # Uncomment if using pytorch-fid instead

# For IS calculation (uses torchvision's InceptionV3)
# torchvision already included above

# Additional evaluation tools
scikit-learn>=1.3.0

# ============================================================================
# Hugging Face (for transformers-based models/verifiers)
# ============================================================================
# Optional: Only needed if using Hugging Face models
transformers>=4.30.0
huggingface-hub>=0.16.0
accelerate>=0.20.0

# ============================================================================
# Visualization and Logging
# ============================================================================
matplotlib>=3.7.0
seaborn>=0.12.0
tensorboard>=2.13.0
wandb>=0.15.0  # Optional: for experiment tracking

# ============================================================================
# Development and Testing
# ============================================================================
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
mypy>=1.5.0  # Optional: for type checking

# ============================================================================
# Additional Utilities
# ============================================================================
requests>=2.31.0  # For downloading models
imageio>=2.31.0  # For image I/O
imageio-ffmpeg>=0.4.3  # For video support (if needed)

# ============================================================================
# Optional: For specific verifiers (from tt-scale-flux)
# ============================================================================
# Uncomment if using these verifiers:
# google-generativeai>=0.3.0  # For Gemini verifier
# openai>=1.0.0  # For OpenAI verifier
# qwen-vl-utils>=0.1.0  # For Qwen verifier (if available)

# ============================================================================
# Notes
# ============================================================================
# 1. EDM model requires code_repos/diffusion-tts/edm/ to be cloned
#    - This provides dnnlib and unet modules
#    - Clone: git clone https://github.com/rvignav/diffusion-tts.git code_repos/diffusion-tts
#
# 2. For ImageNet classifier, the model will be auto-downloaded to ~/.cache/imagenet_classifier/
#
# 3. For EDM pretrained models, they will be downloaded from NVIDIA CDN when first used
#
# 4. Python version: >=3.8, <3.12 (recommended: 3.9 or 3.10)
#
# 5. CUDA: Requires CUDA-capable GPU for model inference
#
# 6. Installation:
#    pip install -r requirements.txt
#
# 7. For development:
#    pip install -r requirements.txt
#    pip install -e .  # Install package in editable mode
